+++ TaskManagerBackend/src/TaskManagerBackend.Application/Features/Tracking/Dtos/TrackingLogEntry/UpdateTrackingLogEntryRequest.cs
@@ -22,7 +22,7 @@ public class UpdateTrackingLogEntryRequest
     public int? Priority { get; init; }
 
     [Range(Constants.MinOrderIndex, Constants.MaxOrderIndex)]
-    public double OrderIndex { get; init; }
+    public decimal OrderIndex { get; init; }
 
     public DateTime UpdatedAt { get; init; }
 }
